Tests check secrets
